requirements.txt
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
backports.pbkdf2==0.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
click==8.0.4
cryptography==36.0.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.1.3
fastapi==0.75.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
Menu==3.2.2
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pycrypto==2.6.1
pycryptodomex==3.14.1
pydantic==1.9.0
pymongo==4.0.2
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.11
uvicorn==0.17.5
